Improve detection of exrm version.
Detect also patch version because this version will be necessary
to decide whether the non-interactive mode should be used
(see next commit).
Also don't check for the locked version because that might be a
git hash, but grep for the version number in the first line of
the `mix deps` output.