Add options for faster and incremental builds.
Allow to override which directories are cleaned before the build is started.
By default `GIT_CLEAN_PATHS` is set to `.` which allows a clean build without
any conflicts, but can be overriden e.g. to keep the deps with something like
"_build rel priv/static/generated". In addition the command line option
`--skip-git-clean` can be used to skip that step completely and start from the
state of the last build and the additional option `--skip-mix-clean` to build
skip the `mix clean` for full incremental builds.

Closes #31