Show command out output when build command failed
even if `--verbose` option is not used. otherwise failed builds
must be rerun with verbose output enabled to be able to fix that
problem. This is achieved by executing the git and build command
synchronously which enables us to capture the output and display
it if the command failed no matter whether verbose output was
enabled or not.