Validate that the version of the built upgrade differs from the original version
the upgrade is built from. And validate this already in the `build upgrade` task,
not in the `upgrade` task (which uses it).
These releases cannot be deployed as upgrades anyway and the `build upgrade`
taks will already fail when detecting the built release, because it is moved
when actually the old / original release is (re)moved and the built release
will not be found for copying.