Don't accept new connections / requests during phoenix app upgrades.
Implement relup instructions which suspends all ranch acceptors
during the upgrade and resumes them when the upgrade is finished.
Before executing the upgrade it is checked whether the ranch
acceptors can be suspended by checking whether they can be
found in the supervision tree of the application. If not, the
upgrade is canceled and the running node is not touched.