Don't consider load-instructions for `RunnableInstruction`s as code loading instructions for the release
when appending instructions to the existing instructions after to point
of no return and before any code, processes or applications of the running
release are modified. These instructions are inserted by the
`RunnableInstruction` itself in `ensure_module_loaded_before_instruction/2`
to ensure that the (new) code is loaded before the instruction is executed
during the upgrade process.