Don't use stdin when edeliver is executed by the mix task.
Opening stdin when ssh command is used with the `-t` option
(pseudo tty allocation) will cause ssh to print the following warning:

 `Pseudo-terminal will not be allocated because stdin is not a terminal.`

It behaves the same way as if stdin would be passed directly to the ssh
command like this:

```sh
  $  ssh -t  user@host 'echo foo | read a; echo $a'
  foo
  Connection to host closed.
  $ echo bar | ssh -t  user@host 'read a; echo $a'
  Pseudo-terminal will not be allocated because stdin is not a terminal.
  bar
```

To avoid this behaviour the invocation of the edeliver script by the
`$Mix.shell.cmd/1` function is replaced by opening a port directly
and using the `:nouse_stdio` option.

This fixes #42.