Monitor process of mix task which started edeliver.
This is required because edeliver is not terminated automatically
when started by the mix task and the mix task is terminated by
`ctrl+c`. To be able to still terminate and close all open ssh
connection that process is monitored and edeliver is terminated
too when the mix task was aborted.