Stop and start the node instead of calling restart
The restart command restarts the application within the same running VM with the same boot args. This means that it won't pick up any newly deployed code. There is a reboot command that will do what we want, but only if heart is running. The best way, for right now, is to stop and then start the node.