Don't suspend the `ranch_acceptors_sup` during upgrades
because suspending a supervisor prevents the usage of `suspend`,
`code_change` and `resume` relup instructions later in the
relup file because the supervison tree is used to find processes
using changed calback modules. Since no acceptors are started
dynamically this can be ignored and suspending that supervisor
is removed. See #68.