Allow to monitor startup progress
by adding rpc call that waits until an edeliver process is started.
This requires that edeliver runs as application and is started as
last application of the release.

If verbose option is enabled the progress is printed to the console
by temporarily adding an additional `:error_logger` report handler
which prints all started applications.