Use configured `RELEASE_DIR` when building upgrades.
The `RELEASE_DIR` can't be autodetected when building upgrades and
using exrm as release tool, because the old release must be extracted
to the correct release directory before builing and auto-detecting the
release dir is possible only after builing. When using umbrella projects
the release dir is not `${BUILD_AT}/rel/${APP}` but
`$BUILD_AT/apps/$APP/rel/$APP` so we need to be able to configure that
in the `.deliver/config` and use the configured value instead. See #88