Pass the branch which is built to the auto-versioning mix task.
This is required since auto-detecting the branch which contains the
current commit/revision to build might result in serveral results,
e.g. if the commit was built in a feature branch and is later merged.
Then the commit would exist in both branches on the build host and
`git branch --contains <revision>` returns two results. Because of
that we use the values passed to the `--branch=` edeliver option
if it is available and the branch detection as fallback if the
auto-versioning feature is used locally. See #95 and #74.