Don't use exrm as dependency if distillery is used
as release tool. This is required because both dependencies
define a `release` and `release.clean` task which results in
duplicate definitions of the `Mix.Tasks.Release` and
`Mix.Tasks.Release.Clean` module.
Not requiring exrm as dependency results in missing base modules
for edeliver exrm plugins, that's why they must be omitted in the
build path then.
In addition the config passed to edeliver relup instruction won't
be a `ReleaseManager.Config` struct, but a `Mix.Releases.Config.t`
struct.