Use only a single distillery plugin for linking configs.
Use same plugin to link `vm.args` and `sys.config` because they
need to be specified in the distillery config as plugin.
Another reason link both in the same module is, that the release
tar needs to be repackaged, becaues distillery uses
`:systools_make.make_tar(...)` to create the release tar which
resoves all links using the `:dereference` option.