Link config files only in release tar
not in release dir. This allows to use links to
destinations which don't need to exist on the build
host. Linking the files in the release dir also causes
distillery to validate their contents (which might fail
if the destination is only a dummy file) and to fail on
the second build, because distillery cannot overwrite
the existing links.