Run epmd-less and use edeliver epmd module
allowing to start a remote console which connects to the running
node and to connect to other nodes, even if they run on the same
host but on a different dist port. In that the known nodes and
its ports can be passed as `edelvier -> nodes` application config
in the sys.config or as EDELIVER_NODES env var.