Set used release command as release image label
and use it when deploying the default docker start script
provided by edeliver. This script is used if no custom container
start script is embedded into the releae image. The default start
script needs to know whether distillery or rebar3 was used to built
the release because their extended start scripts require different
values to be set for the ERL_DIST_PORT env when starting a remote console.
rebar3 needs the same value used when starting the release while
distillery needs a different port because it binds also the remote
console on that port.