Set a temporary workdir while copying built release into docker image
If the release base image set as `DOCKER_RELEASE_BASE_IMAGE`,
contains a workdir set to the application dir
 - which makes sense - e.g. `WORKDIR /my-app` and the built release
 is copied with `docker cp` to `release-container:/my-app`, it will end up
 unexpectedly in `/my-app/my-app` - which is not the goal.
 According [to the docs](https://docs.docker.com/engine/reference/builder/#workdir):

> The WORKDIR instruction sets the working directory for any RUN, CMD, ENTRYPOINT, COPY and ADD instructions

and the unexpected result seen if the `WorkingDir` is set in the container config,
we can avoid copying to the workdir as subdir if we set another temporary workdir
while copying. If this is a different dir (e.g. `/tmp`), the release won't end up
in `/tmp/my-app` but in the expected `/my-app` root destination.

Unfortunately `docker commit` would also commit the temporary workdir,
so it needs to be reset then. If it was not set, it is set to `/$APP` as deault.

Steps to reproduce:

```sh
$ mkdir -p /tmp/foo
$ echo hello > /tmp/foo/hello
$ container_id="$(docker create --workdir=/foo alpine tail -f /dev/null)"
$ docker cp /tmp/foo ${container_id}:/foo
$ docker start $container_id
04ebd37f3a099bdd8893849a785b68df9b26a4328270b45974b898b387d5f24e
$ docker exec -ti $container_id ls -al /foo/
total 12
drwxr-xr-x    3 root     root          4096 Jun 22 12:59 .
drwxr-xr-x    1 root     root          4096 Jun 22 12:59 ..
drwxr-xr-x    2 501      root          4096 Jun 22 12:59 foo
```